mmc: add partboot command
In order to decide which kernel and rootfs to boot we need to know
what was the boot partition for U-Boot.

With "mmc partconf" PARTITION_CONFIG value is displayed but I didn't
any way to use it in the boot script.

Hence, add new "mmc partboot" command that returns true if the
specified boot partition is the current one, otherwise it returns
false.

Upstream-Status: Pending [not yet submitted to upstream]
Signed-off-by: Grygorii Tertychnyi <grygorii.tertychnyi@leica-geosystems.com>