imx8mm-var-dart: config: cleanup environment
As the environment of variscite module is intended to provide a user
with all kind of possibilities to boot, it contains multiple variables
we never use.

The current reserved space of 4K for the environment is filled up to
3564 bytes with mainly variables we never use.

Reduce this to the variables we actually need for our boot process (load
and run bootscript). With this we use only 1171 bytes of the 4K.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>