Revert "configs: imx8mm-ap20: decrease bootdelay to 0"
This reverts commit f985fb53b347b1f87441ed0fb6146d9ac20f1675.

In order to recognise our 'lll' boot sequence, we cannot have a bootdelay
of 0 seconds, because the autoboot process is implemented as a do-while
loop that, with CONFIG_BOOTDELAY=0, terminates after reading just one char.
So, it is impossible to combine CONFIG_BOOTDELAY=0 and
CONFIG_AUTOBOOT_KEYED with CONFIG_AUTOBOOT_STOP_STR longer than 1.

https://github.com/hexagon-geo-surv/u-boot-imx/blob/3dece3a1c1c02b257ae2a2e3c7d770a3fb3949bb/common/autoboot.c#L190

We restore CONFIG_BOOTDELAY=1 so that we can read 'lll' and we are still
fast in booting the device.

Signed-off-by: Massimo Toscanelli <massimo.toscanelli@leica-geosystems.com>