imx8mm_var_dart.h: move board_id logic
Substitute the fit_config definition with a get_fit_config command
that can be called by the boot script (boot from eMMC) or
CONFIG_BOOTCOMMAND (boot from QSPI).

This allows to abstract the fetch of the fit_config in the boot script,
that can be reused for multiple devices and, at the same time, to
bring the same logic in the QSPI image.

Modified also qspiboot command to boot with the preselected fit_config.

Signed-off-by: Massimo Toscanelli <massimo.toscanelli@leica-geosystems.com>