common: leica-sep: terminate any possible characters sent to MCU
We can not guarantee, that a possible transition on TX line of the UART
connection was interpreted as the start of a command by MCU.
Terminate anything that was sent before by sending termination
characters before the command.
With this we make sure that we do not extend anything that was sent
before with our command and MCU interprets `<chunk><cmd>`.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>