common: leica_sep: empty RX FIFO before sending command
If we have a full RX FIFO, we read out and interpret what was received
before we even sent the command. This can lead to missing characters of
the response to our command.

Empty the RX FIFO before sending our command in order to interpret only
packets, which was received after we sent the command.

Signed-off-by: Thomas Haemmerle <thomas.haemmerle@leica-geosystems.com>